♻️ refactor(#56): libs/ 내부의 엔티티 관련 폴더 이름 entities로 수정
