+++ src/api/apps/looks/respositories/i-look.app.repository.interface.ts
@@ -1,5 +1,5 @@
 import { IRepository } from '@src/common/interfaces/i-repository.interface';
-import { LookEntity } from '@src/libs/looks/entity/look.entity';
+import { LookEntity } from '@src/libs/looks/entities/look.entity';
 
 export interface ILookAppRepository extends IRepository<LookEntity> {
   findAll(): Promise<LookEntity[]>;
