+++ src/libs/looks/entities/look.entity.ts
